requirements.txt
smbprotocol
httpx
pathlib
numpy
camelot-py
PyMuPDF
opencv-python-headless
tabula-py
ghostscript
pydantic
PyPDF2==2.12.1
pdfminer.six
pdfminer.six[image]